Client Game View with GUI and card interface
Most of the componenents to get the client to work are here. I have created a "Generic" card game object as a place holder as well as the assets we are using on the Host end. I need some help combining what i have from the placeholder into the 3D object card.